scripts/requirements-modified.txt
requests==2.31.0
python-dotenv==1.0.0
scholarly==1.7.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
uuid==1.30
ratelimit==2.2.1 
